ProcessBuilder java 7 dependent code. Ready for 1.6.2!
Added some java 7 dependent code. Hopefully this isnt a problem and
hopefully only depends on the jdk. Also, I think we are ready for 1.6.2!
Just need to update the download links once the mods are updated and we
are ready to merge!

Note to self: Change tags to include minecraft version also; so the tag
when this branch is done will be v-0.3_MC-1.6.2